Factor refcounting/releasing logic into own class
