+++ x-pack/plugin/esql/compute/src/main/java/org/elasticsearch/compute/data/AbstractBlock.java
@@ -11,8 +11,7 @@ import org.elasticsearch.core.Nullable;
 
 import java.util.BitSet;
 
-abstract class AbstractBlock implements Block {
-    private int references = 1;
+abstract class AbstractBlock extends AbstractThreadLocalRefCounted implements Block {
     private final int positionCount;
 
     @Nullable
@@ -104,52 +103,4 @@ abstract class AbstractBlock implements Block {
     public final boolean isReleased() {
         return hasReferences() == false;
     }
-
-    @Override
-    public final void incRef() {
-        if (isReleased()) {
-            throw new IllegalStateException("can't increase refCount on already released block [" + this + "]");
-        }
-        references++;
-    }
-
-    @Override
-    public final boolean tryIncRef() {
-        if (isReleased()) {
-            return false;
-        }
-        references++;
-        return true;
-    }
-
-    @Override
-    public final boolean decRef() {
-        if (isReleased()) {
-            throw new IllegalStateException("can't release already released block [" + this + "]");
-        }
-
-        references--;
-
-        if (references <= 0) {
-            closeInternal();
-            return true;
-        }
-        return false;
-    }
-
-    @Override
-    public final boolean hasReferences() {
-        return references >= 1;
-    }
-
-    @Override
-    public final void close() {
-        decRef();
-    }
-
-    /**
-     * This is called when the number of references reaches zero.
-     * It must release any resources held by the block (adjusting circuit breakers if needed).
-     */
-    protected abstract void closeInternal();
 }
